refactor : 이름 검증 Name 클래스 내부로 이동
